t2/requirements.txt
pandas
kafka-python
pyyaml>=6.0.0
psycopg2-binary